requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
filelock==3.9.0
fsspec==2023.10.0
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
nltk==3.8.1
numpy==1.24.1
packaging==23.2
pdfkit==1.0.0
Pillow==9.3.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.1
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.14.1
transformers==4.35.0
typing_extensions==4.4.0
urllib3==1.26.13
